Finished simple modeling of the library's data There are four basic models - GutenbergBook - OclcBook - OclcWork - Equivalency
None of these models have an auto incrementing ID or an ID column.
They each have an NID column(native id), which is the ID the library
provided us with, and serves as their primary key. Considering Rails's
convention is to have an ID column that serves as the primary key and
foreign key for relationships, there was some significant amount of
code I needed to add to establish ActiveRecord's relationship mechanism
successfully on these models.

Rub rake db:reset (which should also seed your db with the library's
data) then go to the console and get familiar with the data if you
plan to write some code with it.