Created   - EnumMap class (lib/enum_map.rb)   - EnumHandler concern (app/models/concerns/enum_handler.rb)
The EnumHandler module overides ActiveRecord#enum to recieve a code
block. Which is used to define the helper methods that enum typically
defines for you. The EnumMap class is responsible for keeping track
of the original enum values and allowing you to get/set them as you
would normally, while having methods that are named differently.